Feature: Managed Assets
This changeset introduces a managed asset layer to Elide, which is
integrated into the framework build tools. At build time, embedded
SSR and CSR assets are packaged into the application, and are now
additionally embedded with an asset descriptor providing metadata
and dependency info.

These descriptors are then used at runtime to more efficiently
serve static assets directly out of memory, and to reference static
assets easily from dynamic contexts.

Later, we can use this as groundwork to enable dependency-aware
asset serving, on-the-fly compilation, and so on.

Changes enclosed:
- Add `AssetManager`, `AssetReader`, `AssetResolver` interfaces
- Add initial abstract logic for resolving and serving assets
- Add server configuration properties for the asset system
- Add concept of general bundled asset to `bundle.proto`
- Add built-in implementations for each asset component interface
- Add initial implementation of controller for asset serving

Changes pending:
- Logic to translate assets into HTTP responses
- Logic to load and read assets
- Tests for asset components
- Tests for asset controller
- Gradle plugin support