Feature: SSR Shared State
This changeset introduces the concept of shared state, which is
computed during a `PageController` execution, and shared with the
render context.

State is shared with the frontend via an injected JSON blob which
the framework can decode on behalf of the developer, similar to how
`getServerSideProps` works in Next.js.

VM-side, state is exported for guest language use, allowing access
to context/state within SSR bundles.

Changes enclosed:
- Introduce new `PageWithProps` controller, and `props()` extension
  point, where the developer provides shared state for a given cycle
- Support basic for state sharing with SSR scripts
- Serialize and inject server-side props for frontend use
- Provide `boot` entrypoint for frontend apps, which accounts for
  server-injected state, + SSR vs. CSR mode
- Fix for double-render / double-compute of state
- Fix support for VM inspection via `-Delide.vm.inspect`
- Fix for `runBlocking` call during HTML rendering